A small refactor to simplify and clarify the code.
